+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # 1.0.0 (2021-05-31)
 
+
+### Bug Fixes
+
+* string to number ([9696072](https://github.com/Cansiny0320/qq-group-bot/commit/9696072a8cb6ce344d6605580655b9bfc0461ff7))
+
+
 ### Features
 
-- 群员撤回消息提示 c531004
-- 群员修改昵称提示 d4ea042
+* 群员撤回消息提示 ([c531004](https://github.com/Cansiny0320/qq-group-bot/commit/c531004ab4224fe3621be77bcdf3bebd2a57323c))
+* 群员修改昵称提示 ([d4ea042](https://github.com/Cansiny0320/qq-group-bot/commit/d4ea04283bcaafdd10d9f5dcd4d619f026b1342a))
+
+
+
+